% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_link.R
\name{climate_link}
\alias{climate_link}
\title{Retrieve climate data from a single source}
\usage{
climate_link(climatologia_link)
}
\arguments{
\item{climatologia_link}{Character. URL of the climatology webpage.}
}
\value{
A tibble containing monthly values of temperature and precipitation for the specified location.
}
\description{
Scrapes monthly climatological data from a single Climatempo webpage.
}
\details{
Main function to retrieve climate link
}
\keyword{internal}
