% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-linksdataset.R
\docType{data}
\name{linksdataset}
\alias{linksdataset}
\title{Climate links dataset}
\format{
A data frame with 5 variables:
\describe{
\item{id}{Location ID}
\item{region}{Region name}
\item{state}{State abbreviation}
\item{city_name}{City name}
\item{links}{Climatempo URL}
}
}
\usage{
linksdataset
}
\description{
A dataset containing links and metadata for Brazilian municipalities
}
\keyword{datasets}
