% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_region.R
\name{match_region}
\alias{match_region}
\title{Match region input}
\usage{
match_region(region_input)
}
\arguments{
\item{region_input}{Character vector containing user-provided region names.}
}
\value{
A character vector of matched region names.
}
\description{
Matches user-provided input to valid Brazilian regions using flexible and approximate string matching.
}
\details{
Main function to retrieve match_region
}
\keyword{internal}
