% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_state.R
\name{match_state}
\alias{match_state}
\title{Match state input}
\usage{
match_state(state_input)
}
\arguments{
\item{state_input}{Character vector containing user-provided state names or abbreviations.}
}
\value{
A character vector of matched state abbreviations.
}
\description{
Matches user-provided input to valid Brazilian state abbreviations using flexible and approximate string matching.
}
\details{
Main function to retrieve match_state
}
\keyword{internal}
