% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_normalize_text.R
\name{utils_normalize_text}
\alias{utils_normalize_text}
\title{Normalize text}
\usage{
utils_normalize_text(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A normalized character vector.
}
\description{
Normalizes text by converting to lowercase, removing diacritics, and eliminating whitespace.
}
\details{
Main function to retrieve utils_normalize_text
}
\keyword{internal}
